If we find a link pointing to ., make a similar link to ., not a link to an absolute name.
If we find a link pointing to ., make a similar
link to ., not a link to an absolute name.
(stdio.h): Change stdio.h to use stdarg.h to get va_list.

When link points to a dir $dirname outside $INPUT,
copy that dir into $LIB/root$dirname.  Translate the link itself
into a link to that new dir.  Make a file DONE in each treetop dir
so no treetop is scanned twice.  Improve progress messages.

From-SVN: r2046