Move TARGET_FILE_SWITCHING mips.h hack to psx.h
In the imported mips.h from Psy-Q 4.4 fork of GCC, TARGET_FILE_SWITCHING
option was hacked to be enabled. This option puts all externs at the 
beginning of the generated .asm files. A cleaner way to override this
flag is to change it in psx.h file, not the general mips.h file.