Add UNIX line ending to DOS line ending conversion
This commit adds a new flag: "-fdos-line-endings" (turned on by default)
that causes the compiler to emit the assembler code with DOS line
endings instead of default UNIX line endings (as was the case before).

This is required for compatiblity with original PsyQ utilites (such as
ASPSX.EXE).

The entire GCC code relies on a global asm_out_file variable to emit the
assembly. It was not practical to modify each fwrite to it in the
codebase to emit the correct line ending.

Therefore, this code switches asm_out_file to a temporary file and at
the end of compilation it performs the UNIX line ending to DOS line
ending conversion. The output is written to the original asm_out_file.