Add build scripts compiling in Debian with Docker
This change adds a Makefile that builds the project in a Docker
container with Debian 2.1 Slink image. This is actually a more
historically accurate environment, as Debian 2.1 is from 1999 while
the previously recommended Ubuntu VM is from 2006.

The new build system makes it much easier to build the project.

A "old school" naming was introduced to differentiate that build with
the future "modern" builds.

The cc1 binary is now called cc1psx (in line with Psy-Q 4.4 naming)
and is now compiled with "-O2" and "-g" (debug info) flags.

README.md is updated to reflect the new build system and next things
left to do in the project.