Add a GitHub Actions step building mgs_reversing
Add a new GitHub Actions step that downloads the generated Homebrew
Psy-Q package, extracts it and tries to build mgs_reversing
(https://github.com/FoxdieTeam/mgs_reversing) with that compiler.

Proper care must be taken to make sure that the paths are not too long
(exceeding DOS limit), so the entire build happens in /tmp on GitHub
Actions.

Thanks to this job, we now have proper testing of the compiler
inside this repository.

During the build process a original package of Psy-Q 4.4 utilities is
downloaded (from https://github.com/FoxdieTeam/psyq_sdk) and the
original CC1PSX.EXE is replaced with our Homebrew Psy-Q. Only minimal
changes are required in build.py of mgs_reversing to get it working
the same way as the original Psy-Q 4.4.